Move ExecutorError to own module
This change follows the model in other modules where errors existing in
a submodule.
